requirements.txt
python-dotenv==0.21.0
python_telegram_bot==13.14
requests==2.22.0
telegram==0.0.1
tikapi==3.0.7
tqdm==4.62.3
